+++ extensions/ql-vscode/src/cli.ts
@@ -667,15 +667,18 @@ export class CodeQLCliServer implements Disposable {
 
   /**
   * Generate a summary of an evaluation log.
+  * @param endSummaryPath The path to write only the end of query part of the human-readable summary to. 
   * @param inputPath The path of an evaluation event log.
   * @param outputPath The path to write a human-readable summary of it to.
   */
    async generateLogSummary(
     inputPath: string,
     outputPath: string,
+    endSummaryPath: string,
   ): Promise<string> {
     const subcommandArgs = [
       '--format=text',
+      `--end-summary=${endSummaryPath}`,
       inputPath,
       outputPath
     ];
@@ -1279,8 +1282,14 @@ export class CliVersionConstraint {
 
    /**
     * CLI version that supports rotating structured logs to produce one per query.
+    * 
+    * Note that 2.8.4 supports generating the evaluation logs and summaries,
+    * but 2.9.0 includes a new option to produce the end-of-query summary logs to
+    * the query server console. For simplicity we gate all features behind 2.9.0,
+    * but if a user is tied to the 2.8 release, we can enable evaluator logs 
+    * and summaries for them.
     */
-    public static CLI_VERSION_WITH_PER_QUERY_EVAL_LOG = new SemVer('2.8.4');
+    public static CLI_VERSION_WITH_PER_QUERY_EVAL_LOG = new SemVer('2.9.0');
 
   constructor(private readonly cli: CodeQLCliServer) {
     /**/